+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.3.0 (unreleased)
+## 0.3.0 (2020-07-21)
 
 - Updated OR-Tools to 7.7
 - Added `BasicScheduler` class